+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Subgraph Changelog
 
+## 3.3.11 (2023-01-30)
+
+- Index OHM bonds using Gnosis Easy Auctions
+- Include bonds to the circulating/floating supply calculations
+- Remove price-snapshot from protocol-metrics, as it now has its own subgraph
+- Remove redundant subgraph data sources: bonds, sOHM
+- Shift trigger for ProtocolMetrics from stake() function call to LogRebase event, which should increase the speed of indexing
+
 ## 3.2.16 (2023-01-26)
 
 - Amend title of FraxBP staked in Convex and Frax